[ADD] product_packaging_type: Backport from 13.0
