+++ oca_dependencies.txt
@@ -1,4 +1,5 @@
 # list the OCA project dependencies, one per line
 # add a github url if you need a forked version
 web
-server-tools https://github.com/OCA/server-tools
+stock-logistics-barcode
+server-tools
