+++ product_packaging_type/models/__init__.py
@@ -1,1 +1,2 @@
+from . import product_packaging
 from . import product_packaging_type
