+++ product_packaging_type/readme/CONTRIBUTORS.rst
@@ -1,3 +1,7 @@
 * `Camptocamp <https://www.camptocamp.com>`_
 
   * Damien Crier <damien.crier@camptocamp.com>
+
+* `ACSONE SA/NV <https://acsone.eu>`_
+
+  * Laurent Mignon <laurent.mignon@acsone.eu>
