+++ product_packaging_type/views/assets_backend.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <template id="assets_backend" name="product_packaging_type_assets_backend" inherit_id="web.assets_backend">
+        <xpath expr="//script[last()]" position="after">
+            <script type="text/javascript" src="/product_packaging_type/static/src/js/list_view.js"/>
+        </xpath>
+    </template>
+
+</odoo>
